+++ tests/integration/test_main.py
@@ -43,7 +43,6 @@ class TestMain:
             'kytos/.*.link_available_tags',
             'kytos/.*.liveness.(up|down)',
             'kytos/.*.liveness.disabled',
-            'kytos/topology.get',
             '.*.topo_controller.upsert_switch',
             '.*.of_lldp.network_status.updated',
             '.*.interface.is.nni',